+++ encore_api_cli/commands/analyze.py
@@ -2,6 +2,7 @@ import io
 from typing import Optional
 
 import click
+from click_help_colors import HelpColorsGroup
 from yaspin import yaspin
 
 from ..options import common_options
@@ -11,7 +12,7 @@ from ..utils import color_id, get_client, parse_rule
 from .analysis import show
 
 
-@click.group()
+@click.group(cls=HelpColorsGroup, help_options_color="cyan")
 def cli() -> None:  # noqa: D103
     pass