+++ encore_api_cli/commands/download.py
@@ -4,6 +4,7 @@ from typing import Tuple
 from urllib.parse import urlparse
 
 import click
+from click_help_colors import HelpColorsGroup
 
 from ..options import common_options
 from ..output import echo
@@ -11,7 +12,7 @@ from ..state import State, pass_state
 from ..utils import color_path, get_client
 
 
-@click.group()
+@click.group(cls=HelpColorsGroup, help_options_color="cyan")
 def cli() -> None:  # noqa: D103
     pass