+++ encore_api_cli/commands/upload.py
@@ -1,4 +1,5 @@
 import click
+from click_help_colors import HelpColorsGroup
 
 from ..options import common_options
 from ..output import echo_success
@@ -7,7 +8,7 @@ from ..state import State, pass_state
 from ..utils import color_id, color_path, get_client
 
 
-@click.group()
+@click.group(cls=HelpColorsGroup, help_options_color="cyan")
 def cli() -> None:  # noqa: D103
     pass